Add three functions: randomization, timer, fix a bug
(1) Right now, the script presents the stimuli (videos) in the order in which they appear in the stimuli.csv doc. We need them presented in a random order. 
(2) The script breaks after it has displayed the last video. I have no idea why. This needs to be fixed. 
(3) We need to add a countdown sequence ("3...2...1...") before each video. This should be timed across three seconds.